+++ src/endpoint.ts
@@ -70,7 +70,7 @@ export abstract class AbstractEndpoint {
   public abstract getScopes(): string[];
   public abstract getTags(): string[];
   public abstract _setSiblingMethods(methods: Method[]): void;
-  public abstract getOperationId(): string | undefined;
+  public abstract getOperationId(method: Method): string | undefined;
 }
 
 type EndpointProps<
@@ -89,7 +89,7 @@ type EndpointProps<
   resultHandler: ResultHandlerDefinition<POS, NEG>;
   description?: string;
   shortDescription?: string;
-  operationId?: string;
+  operationId?: string | ((method: Method) => string);
 } & ({ method: Method } | { methods: Method[] }) &
   ({ scopes?: SCO[] } | { scope?: SCO }) &
   ({ tags?: TAG[] } | { tag?: TAG });
@@ -119,7 +119,7 @@ export class Endpoint<
   };
   readonly #scopes: SCO[] = [];
   readonly #tags: TAG[] = [];
-  readonly #operationId?: string;
+  readonly #operationId?: string | ((method: Method) => string);
 
   constructor({
     middlewares,
@@ -245,8 +245,10 @@ export class Endpoint<
     return this.#tags;
   }
 
-  public override getOperationId(): string | undefined {
-    return this.#operationId;
+  public override getOperationId(method: Method): string | undefined {
+    return typeof this.#operationId === "function"
+      ? this.#operationId(method)
+      : this.#operationId;
   }
 
   #getDefaultCorsHeaders(): Record<string, string> {